--- a/blosc/blosc2.h
+++ b/blosc/blosc2.h
@@ -676,6 +676,10 @@ static const blosc2_cparams BLOSC2_CPARAMS_DEFAULTS = {
 typedef struct {
   int nthreads;
   //!< The number of threads to use internally (1).
+  bool* block_maskout;
+  //!< The blocks that are not meant to be decompressed. If NULL (default), all blocks in a chunk should be read.
+  int block_maskout_nitems;
+  //!< The number of items in block_maskout array (must match the number of blocks in chunk if block_maskout != NULL).
   void* schunk;
   //!< The associated schunk, if any (NULL).
 } blosc2_dparams;
@@ -683,7 +687,7 @@ typedef struct {
 /**
  * @brief Default struct for decompression params meant for user initialization.
  */
-static const blosc2_dparams BLOSC2_DPARAMS_DEFAULTS = {1, NULL };
+static const blosc2_dparams BLOSC2_DPARAMS_DEFAULTS = {1, NULL, 0, NULL };
 
 /**
  * @brief Create a context for @a *_ctx() compression functions.
@@ -712,6 +716,20 @@ BLOSC_EXPORT blosc2_context* blosc2_create_dctx(blosc2_dparams dparams);
  * both compression and decompression.
  */
 BLOSC_EXPORT void blosc2_free_ctx(blosc2_context* context);
+
+/**
+ * @brief Set a maskout so as to avoid decompressing specified blocks.
+ *
+ * @param dctx The decompression context to update.
+ *
+ * @param block_maskout The boolean mask for avoiding decompression in block.
+ *
+ * @param nblocks The number of blocks in maskout above.
+ *
+ * This function should always succeed and is valid for contexts only meant
+ * for decompression.
+ */
+BLOSC_EXPORT void blosc2_set_maskout(blosc2_context *dctx, bool *block_maskout, int nblocks);
 
 /**
  * @brief Context interface to Blosc compression. This does not require a call