This implements a maskout for reading operations, and allow parallel decompression to work with masks too.