@FrancescAlted 2020-03-26
In this PR, the parallel read algorithm has changed to use a queue for finding new blocks to decompress.  This creates a bit more contention, and decompression speeds do not reach as much as speed as before, but the difference is not too much.

Before the PR:

```
faltet@corex14/tmp/c-blosc2/build $ bench/b2bench blosclz shuffle single 12
Blosc version: 2.0.0-beta.5.dev ($Date:: 2019-09-13 #$)
List of supported compressors in this build: blosclz,lz4,lz4hc,lizard,zlib,zstd
Supported compression libraries:
  BloscLZ: 2.0.0
  LZ4: 1.9.1
  Lizard: 1.0.0
  Snappy: unknown
  Zlib: 10.0.3
  Zstd: 1.4.4
Using compressor: blosclz
Using shuffle type: shuffle
Running suite: single
--> 12, 4194304, 4, 19, blosclz, shuffle
********************** Run info ******************************
Blosc version: 2.0.0-beta.5.dev ($Date:: 2019-09-13 #$)
Using synthetic data with 19 significant bits (out of 32)
Dataset size: 4194304 bytes	Type size: 4 bytes
Working set: 256.0 MB		Number of threads: 12
********************** Running benchmarks *********************
memcpy(write):		  701.1 us, 5705.7 MB/s
memcpy(read):		  418.3 us, 9561.7 MB/s
Compression level: 0
comp(write):	  174.7 us, 22901.1 MB/s	  Final bytes: 4194336  Ratio: 1.00
decomp(read):	  403.9 us, 9903.0 MB/s	  OK
Compression level: 1
comp(write):	  163.5 us, 24459.9 MB/s	  Final bytes: 2125328  Ratio: 1.97
decomp(read):	  108.8 us, 36770.6 MB/s	  OK
Compression level: 2
comp(write):	  145.7 us, 27456.5 MB/s	  Final bytes: 1105168  Ratio: 3.80
decomp(read):	   95.8 us, 41754.0 MB/s	  OK
Compression level: 3
comp(write):	  139.0 us, 28783.9 MB/s	  Final bytes: 1105168  Ratio: 3.80
decomp(read):	   77.1 us, 51861.7 MB/s	  OK
Compression level: 4
comp(write):	  144.1 us, 27761.1 MB/s	  Final bytes: 1085072  Ratio: 3.87
decomp(read):	   87.8 us, 45547.5 MB/s	  OK
Compression level: 5
comp(write):	  222.4 us, 17983.9 MB/s	  Final bytes: 425360  Ratio: 9.86
decomp(read):	  111.6 us, 35838.1 MB/s	  OK
Compression level: 6
comp(write):	  213.3 us, 18750.0 MB/s	  Final bytes: 268224  Ratio: 15.64
decomp(read):	  110.1 us, 36322.3 MB/s	  OK
Compression level: 7
comp(write):	  186.5 us, 21447.1 MB/s	  Final bytes: 268224  Ratio: 15.64
decomp(read):	  105.3 us, 37993.3 MB/s	  OK
Compression level: 8
comp(write):	  182.5 us, 21920.3 MB/s	  Final bytes: 193392  Ratio: 21.69
decomp(read):	  105.6 us, 37890.9 MB/s	  OK
Compression level: 9
comp(write):	  183.4 us, 21808.0 MB/s	  Final bytes: 193392  Ratio: 21.69
decomp(read):	   95.9 us, 41694.6 MB/s	  OK

Round-trip compr/decompr on 7.5 GB
Elapsed time:	    0.8 s, 20624.9 MB/s
```

after PR:

```
faltet@corex14/tmp/c-blosc2/build $ bench/b2bench blosclz shuffle single 12
Blosc version: 2.0.0-beta.5.dev ($Date:: 2019-09-13 #$)
List of supported compressors in this build: blosclz,lz4,lz4hc,lizard,zlib,zstd
Supported compression libraries:
  BloscLZ: 2.0.0
  LZ4: 1.9.1
  Lizard: 1.0.0
  Snappy: unknown
  Zlib: 10.0.3
  Zstd: 1.4.4
Using compressor: blosclz
Using shuffle type: shuffle
Running suite: single
--> 12, 4194304, 4, 19, blosclz, shuffle
********************** Run info ******************************
Blosc version: 2.0.0-beta.5.dev ($Date:: 2019-09-13 #$)
Using synthetic data with 19 significant bits (out of 32)
Dataset size: 4194304 bytes	Type size: 4 bytes
Working set: 256.0 MB		Number of threads: 12
********************** Running benchmarks *********************
memcpy(write):		  710.0 us, 5633.9 MB/s
memcpy(read):		  437.7 us, 9138.9 MB/s
Compression level: 0
comp(write):	  262.1 us, 15261.4 MB/s	  Final bytes: 4194336  Ratio: 1.00
decomp(read):	  418.5 us, 9557.0 MB/s	  OK
Compression level: 1
comp(write):	  168.9 us, 23681.0 MB/s	  Final bytes: 2125328  Ratio: 1.97
decomp(read):	  170.5 us, 23456.8 MB/s	  OK
Compression level: 2
comp(write):	  130.8 us, 30576.0 MB/s	  Final bytes: 1105168  Ratio: 3.80
decomp(read):	  111.4 us, 35906.3 MB/s	  OK
Compression level: 3
comp(write):	  139.4 us, 28687.7 MB/s	  Final bytes: 1105168  Ratio: 3.80
decomp(read):	  107.3 us, 37284.4 MB/s	  OK
Compression level: 4
comp(write):	  137.4 us, 29115.6 MB/s	  Final bytes: 1085072  Ratio: 3.87
decomp(read):	   93.5 us, 42798.6 MB/s	  OK
Compression level: 5
comp(write):	  235.6 us, 16976.3 MB/s	  Final bytes: 425360  Ratio: 9.86
decomp(read):	  103.6 us, 38624.0 MB/s	  OK
Compression level: 6
comp(write):	  194.4 us, 20576.8 MB/s	  Final bytes: 268224  Ratio: 15.64
decomp(read):	   94.7 us, 42238.9 MB/s	  OK
Compression level: 7
comp(write):	  199.2 us, 20079.4 MB/s	  Final bytes: 268224  Ratio: 15.64
decomp(read):	  107.4 us, 37229.6 MB/s	  OK
Compression level: 8
comp(write):	  183.6 us, 21788.7 MB/s	  Final bytes: 193392  Ratio: 21.69
decomp(read):	  100.0 us, 40011.7 MB/s	  OK
Compression level: 9
comp(write):	  190.7 us, 20974.8 MB/s	  Final bytes: 193392  Ratio: 21.69
decomp(read):	  112.7 us, 35492.3 MB/s	  OK

Round-trip compr/decompr on 7.5 GB
Elapsed time:	    0.9 s, 19540.8 MB/s
```