@FrancescAlted 2020-03-26
@oscargm98 could you please confirm that the PR works well for allowind parallel reads with masks in a Caterva sub-partitions context?